fix: It is now impossible to create new components or to open the unit menu while there is no session initiated or imported.
BREAKING CHANGE: Implemented flowtype-heritage during layout configuration. When creating new connections or changing flowtypes of components, the impact on other connections and components are determined and changes on their flowtypes are made if necessary. Components and connections should from now on always have valid flowtype-combinations. The file flowtype_determination.py is introduced as a utility script

feat: Removed the asset-overview-screen as well as the option to create flowtypes via the component creation menu